Lint page imports; reorganize page modules.
