+++ src/pages/About.jsx
@@ -13,12 +13,10 @@ import {
 } from 'semantic-ui-react'
 
 // Components
-import Header from './Header'
-import Hero from './Hero'
-import Footer from './Footer'
-
-import Timeline from '../components/modules/Timeline';
-
+import Header from '../components/modules/Header'
+import Hero from '../components/modules/Hero'
+import Timeline from '../components/modules/Timeline'
+import Footer from '../components/modules/Footer'
 
 // Assets 
 import Misha from '../assets/images/misha-work.jpg'
