+++ src/pages/Advocacy.jsx
@@ -13,11 +13,15 @@ import {
 } from 'semantic-ui-react'
 
 // Components
-import Header from './Header'
-import Hero from './Hero'
+import Header from '../components/modules/Header'
+import Hero from '../components/modules/Hero'
 import Key from '../components/modules/Key'
-import Footer from './Footer'
+import Footer from '../components/modules/Footer'
+
+// Index
 import AdvocacyIndex from '../components/indices/AdvocacyIndex'
+
+// Case Study
 
 // Assets 
 import {
