+++ src/pages/Design.jsx
@@ -12,11 +12,12 @@ import {
 } from 'semantic-ui-react'
 
 // Components
-import Header from './Header'
-import Hero from './Hero'
+import Header from '../components/modules/Header'
+import Hero from '../components/modules/Hero'
 import Key from '../components/modules/Key'
-import Footer from './Footer'
+import Footer from '../components/modules/Footer'
 
+// Index
 import DesignIndex from '../components/indices/DesignIndex'
 
 // Assets
